uncomment ApplicationConfiguration.Initialize();
I'm not sure why this is causing compilation to fail. But without the app seems to lose some of it's theming and some UI elements start looking (to me) like the old Windows Classic theme. I think for users it's better to have the buttons looking and it should be no trouble for you to comment it out again before building it yourself if you need.