Fix for file names with spaces; Form re-sizing, show progress, added file menu (#2)
* Fixes for:
 Form re-sizing
 no progress shown
 filenames with spaces don't work
Added:
 File Menu
 'Clear Items' functionality

* 1. Changed to .NET Framework 4.6 to increase compatibility
2. added Drag & Drop
3. Added context & other menus for lvl list box
4. Added compression icon for app
5. Added Tool Tip for list box
6. Added message for amount reduced for output
7. Added remove item menu item to menus
8. Added about menu

* 1. Removed un-necessary files.
2. restored gui.csproj and deleted extra solution
3. better organized Form1

* Restoring file menu, deleting un-necessary config file

* Back to .net 8

* adding this back with relative path

* small formatting improvement to repackButton_Click

* uncomment ApplicationConfiguration.Initialize();

I'm not sure why this is causing compilation to fail. But without the app seems to lose some of it's theming and some UI elements start looking (to me) like the old Windows Classic theme. I think for users it's better to have the buttons looking and it should be no trouble for you to comment it out again before building it yourself if you need.

---------

Co-authored-by: PrismaticFlower <5582558+PrismaticFlower@users.noreply.github.com>